+++ python/GafferArnoldUI/ArnoldAttributesUI.py
@@ -69,6 +69,25 @@ def __visibilitySummary( plug ) :
 __transformTypeEnumNames = { "linear" : "Linear", "rotate_about_origin" : "RotateAboutOrigin",
 	"rotate_about_center" : "RotateAboutCenter" }
 
+def __autoBumpVisibilitySummary( plug ) :
+
+	info = []
+	for childName, label in (
+
+		( "camera", "Camera" ),
+		( "diffuseReflection", "DiffRefl" ),
+		( "specularReflection", "SpecRefl" ),
+		( "diffuseTransmission", "DiffTrans" ),
+		( "specularTransmission", "SpecTrans" ),
+		( "volume", "Volume" ),
+		( "subsurface", "Subsurf" ),
+
+	)	:
+		if plug[childName+"AutoBumpVisibility"]["enabled"].getValue() :
+			info.append( label + ( " On" if plug[childName+"AutoBumpVisibility"]["value"].getValue() else " Off" ) )
+
+	return ", ".join( info )
+
 def __transformSummary( plug ) :
 
 	info = []
@@ -173,6 +192,7 @@ Gaffer.Metadata.registerNode(
 		"attributes" : [
 
 			"layout:section:Visibility:summary", __visibilitySummary,
+			"layout:section:Displacement.Auto Bump Visibility:summary", __autoBumpVisibilitySummary,
 			"layout:section:Transform:summary", __transformSummary,
 			"layout:section:Shading:summary", __shadingSummary,
 			"layout:section:Subdivision:summary", __subdivisionSummary,
@@ -300,6 +320,111 @@ Gaffer.Metadata.registerNode(
 			""",
 
 			"layout:section", "Visibility",
+			"label", "Subsurface",
+
+		],
+
+		"attributes.autoBump" : [
+
+			"description",
+			"""
+			Automatically turns the details of the displacement map
+			into bump, wherever the mesh is not subdivided enough
+			to properly capture them.
+			""",
+
+			"nodule:type", "",
+			"layout:section", "Displacement",
+
+		],
+
+		"attributes.cameraAutoBumpVisibility" : [
+
+			"description",
+			"""
+			Whether or not the autobump is visible to camera
+			rays.
+			""",
+
+			"layout:section", "Displacement.Auto Bump Visibility",
+			"label", "Camera",
+
+		],
+
+		"attributes.diffuseReflectionAutoBumpVisibility" : [
+
+			"description",
+			"""
+			Whether or not the autobump is visible in
+			reflected diffuse ( ie. if it casts bounce light )
+			""",
+
+			"layout:section", "Displacement.Auto Bump Visibility",
+			"label", "Diffuse Reflection",
+
+		],
+
+		"attributes.specularReflectionAutoBumpVisibility" : [
+
+			"description",
+			"""
+			Whether or not the autobump is visible in
+			reflected specular ( ie. if it is visible in mirrors ).
+			""",
+
+			"layout:section", "Displacement.Auto Bump Visibility",
+			"label", "Specular Reflection",
+
+		],
+
+		"attributes.diffuseTransmissionAutoBumpVisibility" : [
+
+			"description",
+			"""
+			Whether or not the autobump is visible in
+			transmitted diffuse ( ie. if it casts light through leaves ).
+			""",
+
+			"layout:section", "Displacement.Auto Bump Visibility",
+			"label", "Diffuse Transmission",
+
+		],
+
+		"attributes.specularTransmissionAutoBumpVisibility" : [
+
+			"description",
+			"""
+			Whether or not the autobump is visible in
+			refracted specular ( ie. if it can be seen through glass ).
+			""",
+
+			"layout:section", "Displacement.Auto Bump Visibility",
+			"label", "Specular Transmission",
+
+		],
+
+		"attributes.volumeAutoBumpVisibility" : [
+
+			"description",
+			"""
+			Whether or not the autobump is visible in
+			volume scattering.
+			""",
+
+			"layout:section", "Displacement.Auto Bump Visibility",
+			"label", "Volume",
+
+		],
+
+		"attributes.subsurfaceAutoBumpVisibility" : [
+
+			"description",
+			"""
+			Whether or not the autobump is visible to subsurface
+			rays.
+			""",
+
+			"layout:section", "Displacement.Auto Bump Visibility",
 			"label", "Subsurface",
 
 		],